requirements.txt
streamlit
tensorflow
tensorflow-hub
opencv-python-headless
pillow
